+++ app/src/main/java/com/superwall/superapp/test/UITestMocks.kt
@@ -3,25 +3,39 @@ package com.superwall.superapp.test
 import com.superwall.sdk.analytics.superwall.SuperwallEventInfo
 import com.superwall.sdk.delegate.SuperwallDelegate
 import com.superwall.sdk.paywall.presentation.internal.state.PaywallResult
-import com.superwall.sdk.paywall.vc.PaywallViewController
-import com.superwall.sdk.paywall.vc.delegate.PaywallViewControllerDelegate
+import com.superwall.sdk.paywall.vc.PaywallView
+import com.superwall.sdk.paywall.vc.delegate.PaywallViewCallback
 
-class MockPaywallViewControllerDelegate : PaywallViewControllerDelegate {
-    private var paywallViewControllerDidFinish: ((PaywallViewController, PaywallResult, Boolean) -> Unit)? = null
+@Deprecated("Will be removed in the upcoming versions, use MockPaywallViewDelegate instead")
+typealias MockPaywallViewControllerDelegate = MockPaywallViewDelegate
+
+class MockPaywallViewDelegate : PaywallViewCallback {
+    private var paywallViewFinished: ((PaywallView, PaywallResult, Boolean) -> Unit)? = null
 
     override fun didFinish(
-        paywall: PaywallViewController,
+        paywall: PaywallView,
+        result: PaywallResult,
+        shouldDismiss: Boolean,
+    ) = onFinished(paywall, result, shouldDismiss)
+
+    override fun onFinished(
+        paywall: PaywallView,
         result: PaywallResult,
         shouldDismiss: Boolean,
     ) {
-        paywallViewControllerDidFinish?.invoke(paywall, result, shouldDismiss)
+        paywallViewFinished?.invoke(paywall, result, shouldDismiss)
         if (shouldDismiss) {
             paywall.encapsulatingActivity?.finish()
         }
     }
 
-    fun paywallViewControllerDidFinish(handler: (PaywallViewController, PaywallResult, Boolean) -> Unit) {
-        paywallViewControllerDidFinish = handler
+    @Deprecated("Will be removed in the upcoming versions, use paywallViewFinished instead")
+    fun paywallViewControllerDidFinish(handler: (PaywallView, PaywallResult, Boolean) -> Unit) {
+        paywallViewFinished(handler)
+    }
+
+    fun paywallViewFinished(handler: (PaywallView, PaywallResult, Boolean) -> Unit) {
+        paywallViewFinished = handler
     }
 }