+++ superwall/src/main/java/com/superwall/sdk/Superwall.kt
@@ -24,12 +24,12 @@ import com.superwall.sdk.paywall.presentation.PaywallInfo
 import com.superwall.sdk.paywall.presentation.PresentationItems
 import com.superwall.sdk.paywall.presentation.internal.dismiss
 import com.superwall.sdk.paywall.presentation.internal.state.PaywallResult
-import com.superwall.sdk.paywall.vc.PaywallViewController
+import com.superwall.sdk.paywall.vc.PaywallView
 import com.superwall.sdk.paywall.vc.SuperwallPaywallActivity
 import com.superwall.sdk.paywall.vc.SuperwallStoreOwner
 import com.superwall.sdk.paywall.vc.ViewModelFactory
 import com.superwall.sdk.paywall.vc.ViewStorageViewModel
-import com.superwall.sdk.paywall.vc.delegate.PaywallViewControllerEventDelegate
+import com.superwall.sdk.paywall.vc.delegate.PaywallViewEventCallback
 import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent
 import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent.Closed
 import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent.Custom
@@ -37,7 +37,7 @@ import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent.InitiateP
 import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent.InitiateRestore
 import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent.OpenedDeepLink
 import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent.OpenedURL
-import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent.OpenedUrlInSafari
+import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent.OpenedUrlInChrome
 import com.superwall.sdk.storage.ActiveSubscriptionStatus
 import com.superwall.sdk.store.ExternalNativePurchaseController
 import kotlinx.coroutines.CoroutineScope
@@ -60,7 +60,7 @@ class Superwall(
     options: SuperwallOptions?,
     private var activityProvider: ActivityProvider?,
     private val completion: (() -> Unit)?,
-) : PaywallViewControllerEventDelegate {
+) : PaywallViewEventCallback {
     private var _options: SuperwallOptions? = options
     private val ioScope = CoroutineScope(Dispatchers.IO)
     internal var context: Context = context.applicationContext
@@ -79,10 +79,10 @@ class Superwall(
     internal val presentationItems: PresentationItems = PresentationItems()
 
     /**
-     * The presented paywall view controller.
+     * The presented paywall view.
      */
-    val paywallViewController: PaywallViewController?
-        get() = dependencyContainer.paywallManager.presentedViewController
+    val paywallView: PaywallView?
+        get() = dependencyContainer.paywallManager.currentView
 
     /**
      * A convenience variable to access and change the paywall options that you passed
@@ -104,7 +104,7 @@ class Superwall(
      * Determines whether a paywall is being presented.
      */
     val isPaywallPresented: Boolean
-        get() = paywallViewController != null
+        get() = paywallView != null
 
     /**
      * The delegate that handles Superwall lifecycle events.
@@ -182,12 +182,12 @@ class Superwall(
         get() = dependencyContainer.identityManager.isLoggedIn
 
     /**
-     * The `PaywallInfo` object of the most recently presented view controller.
+     * The `PaywallInfo` object of the most recently presented view.
      */
     val latestPaywallInfo: PaywallInfo?
         get() {
             val presentedPaywallInfo =
-                dependencyContainer.paywallManager.presentedViewController?.info
+                dependencyContainer.paywallManager.currentView?.info
             return presentedPaywallInfo ?: presentationItems.paywallInfo
         }
 
@@ -373,9 +373,9 @@ class Superwall(
      */
     fun togglePaywallSpinner(isHidden: Boolean) {
         ioScope.launch {
-            val paywallViewController =
-                dependencyContainer.paywallManager.presentedViewController ?: return@launch
-            paywallViewController.togglePaywallSpinner(isHidden)
+            val paywallView =
+                dependencyContainer.paywallManager.currentView ?: return@launch
+            paywallView.togglePaywallSpinner(isHidden)
         }
     }
 
@@ -486,12 +486,12 @@ class Superwall(
 
     override suspend fun eventDidOccur(
         paywallEvent: PaywallWebEvent,
-        paywallViewController: PaywallViewController,
+        paywallView: PaywallView,
     ) {
         withContext(Dispatchers.Main) {
             Logger.debug(
                 logLevel = LogLevel.debug,
-                scope = LogScope.paywallViewController,
+                scope = LogScope.paywallView,
                 message = "Event Did Occur",
                 info = mapOf("event" to paywallEvent),
             )
@@ -499,7 +499,7 @@ class Superwall(
             when (paywallEvent) {
                 is Closed -> {
                     dismiss(
-                        paywallViewController,
+                        paywallView,
                         result = PaywallResult.Declined(),
                         closeReason = PaywallCloseReason.ManualClose,
                     )
@@ -515,7 +515,7 @@ class Superwall(
                             try {
                                 dependencyContainer.transactionManager.purchase(
                                     paywallEvent.productId,
-                                    paywallViewController,
+                                    paywallView,
                                 )
                             } finally {
                                 // Ensure the task is cleared once the purchase is complete or if an error occurs
@@ -525,14 +525,14 @@ class Superwall(
                 }
 
                 is InitiateRestore -> {
-                    dependencyContainer.transactionManager.tryToRestore(paywallViewController)
+                    dependencyContainer.transactionManager.tryToRestore(paywallView)
                 }
 
                 is OpenedURL -> {
                     dependencyContainer.delegateAdapter.paywallWillOpenURL(url = paywallEvent.url)
                 }
 
-                is OpenedUrlInSafari -> {
+                is OpenedUrlInChrome -> {
                     dependencyContainer.delegateAdapter.paywallWillOpenURL(url = paywallEvent.url)
                 }