+++ superwall/src/main/java/com/superwall/sdk/paywall/vc/web_view/messaging/PaywallWebEvent.kt
@@ -26,7 +26,7 @@ sealed class PaywallWebEvent {
     ) : PaywallWebEvent()
 
     @SerialName("opened_url_in_safari")
-    data class OpenedUrlInSafari(
+    data class OpenedUrlInChrome(
         val url: URL,
     ) : PaywallWebEvent()